remove decimal pixels, move functions to hooks, svgs to _svg folder, constants to const folder,
