+++ src/app/murmur/dashboard/useDashboard.tsx
@@ -4,7 +4,7 @@ import { useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { z } from 'zod';
 import { EmailVerificationStatus, UserContactList } from '@prisma/client';
-import { useEffect, useState, useMemo, useCallback } from 'react';
+import { useEffect, useState, useMemo, useCallback, FormEvent } from 'react';
 import { useCreateCampaign } from '@/hooks/queryHooks/useCampaigns';
 import { urls } from '@/constants/urls';
 import {
@@ -23,6 +23,14 @@ import { TableCellTooltip } from '@/components/molecules/TableCellTooltip/TableC
 import { useMe } from '@/hooks/useMe';
 import { StripeSubscriptionStatus } from '@/types';
 import { usePageTransition } from '@/contexts/PageTransitionContext';
+import {
+	canadianProvinceAbbreviations,
+	canadianProvinceNames,
+	stateBadgeColorMap,
+} from '@/constants';
+import { useClerk } from '@clerk/nextjs';
+import { useIsMobile } from '@/hooks/useIsMobile';
+import { isProblematicBrowser } from '@/utils/browserDetection';
 
 const formSchema = z.object({
 	searchText: z.string().min(1, 'Search text is required'),
@@ -38,88 +46,6 @@ export const useDashboard = () => {
 	const MAX_CELL_LENGTH = 35;
 
 	// Background color map for state badges (to be filled with provided hex codes)
-	const stateBadgeColorMap = useMemo(
-		() =>
-			({
-				AL: '#F3D7D7',
-				AK: '#D7D7F3',
-				AZ: '#E7F307',
-				AR: '#F3D7F0',
-				CA: '#D7F3EE',
-				CO: '#F3E6D7',
-				CT: '#DEB7F3',
-				DE: '#DBF3D7',
-				FL: '#F3D7E0',
-				GA: '#D7F3F3',
-				HI: '#F1B7F3',
-				ID: '#EDF7F3',
-				IL: '#D7F3E5',
-				IN: '#F3DDD7',
-				IA: '#D7D9F3',
-				KS: '#E2F3D7',
-				KY: '#F3DFA2',
-				LA: '#D7F3F3',
-				ME: '#F3ECD7',
-				MD: '#EDF7F3',
-				MA: '#D7F3DC',
-				MI: '#F3D7D8',
-				MN: '#D7F3E3',
-				MS: '#EBF307',
-				MO: '#F3D7F3',
-				MT: '#D7F3EB',
-				NE: '#F3EBD7',
-				NV: '#DAD7F3',
-				NH: '#DCF3D7',
-				NJ: '#DCF3D7',
-				NM: '#DCF3C7',
-				NY: '#F3F2D7',
-				NC: '#EAD7F3',
-				ND: '#D7F3E1',
-				OH: '#F3D9D7',
-				OK: '#D0F3D7',
-				OR: '#E5F3D7',
-				PA: '#F3D7ED',
-				RI: '#D7F3F1',
-				SC: '#F3E8D7',
-				SD: '#E0F7F3',
-				TN: '#D7F3B8',
-				TX: '#F3D7DE',
-				UT: '#D7E6F3',
-				VT: '#EFF3D7',
-				VA: '#EDF7F3',
-				WA: '#D7F3E7',
-				WV: '#F3DFD7',
-				WI: '#D7F3F3',
-				WY: '#DFF307',
-				// Note: DC not provided; add later if needed
-				// Unspecified states default to transparent background
-			} as Record<string, string>),
-		[]
-	);
-
-	// Canadian provinces detection (by full name and abbreviation)
-	const canadianProvinceNames = useMemo(
-		() =>
-			new Set(
-				[
-					'Alberta',
-					'British Columbia',
-					'Manitoba',
-					'New Brunswick',
-					'Newfoundland and Labrador',
-					'Nova Scotia',
-					'Ontario',
-					'Prince Edward Island',
-					'Quebec',
-					'Saskatchewan',
-				].map((s) => s.toLowerCase())
-			),
-		[]
-	);
-	const canadianProvinceAbbreviations = useMemo(
-		() => new Set(['AB', 'BC', 'MB', 'NB', 'NL', 'NS', 'ON', 'PE', 'QC', 'SK']),
-		[]
-	);
 
 	type TabValue = 'search' | 'list';
 	type TabOption = {
@@ -148,8 +74,10 @@ export const useDashboard = () => {
 	});
 
 	/* HOOKS */
-	// useMe will return undefined values for unauthenticated users
 	const { isFreeTrial, user } = useMe() || { isFreeTrial: false, user: null };
+	const { isSignedIn, openSignIn } = useClerk();
+	const hasProblematicBrowser = isProblematicBrowser();
+	const isMobile = useIsMobile();
 
 	// User can search if they have an active subscription OR are on a free trial
 	const canSearch =
@@ -281,6 +209,23 @@ export const useDashboard = () => {
 		setLimit(50);
 		setHasSearched(true);
 		// The query will automatically run when the state updates enable it
+	};
+
+	const handleSubmitSearch = (e: FormEvent<HTMLFormElement>) => {
+		e.preventDefault();
+		if (!isSignedIn) {
+			if (hasProblematicBrowser) {
+				console.log('[Dashboard] Edge/Safari detected, navigating to sign-in page');
+				if (typeof window !== 'undefined') {
+					sessionStorage.setItem('redirectAfterSignIn', window.location.pathname);
+				}
+				window.location.href = urls.signIn.index;
+			} else {
+				openSignIn();
+			}
+		} else {
+			form.handleSubmit(onSubmit)(e);
+		}
 	};
 
 	const handleResetSearch = () => {
@@ -482,7 +427,7 @@ export const useDashboard = () => {
 			},
 			{
 				accessorKey: 'title',
-				size: 250, // Even width distribution
+				size: 250,
 				header: () => (
 					<span className="font-medium font-secondary text-[14px] ml-2">Title</span>
 				),
@@ -510,7 +455,7 @@ export const useDashboard = () => {
 			},
 			{
 				id: 'place',
-				size: 180, // Even width distribution
+				size: 180,
 				header: () => (
 					<span className="font-medium font-secondary text-[14px]">Place</span>
 				),
@@ -590,7 +535,7 @@ export const useDashboard = () => {
 			},
 			{
 				accessorKey: 'email',
-				size: 280, // Even width distribution
+				size: 280,
 				header: () => (
 					<span className="font-medium font-secondary text-[14px]">Email</span>
 				),
@@ -606,14 +551,7 @@ export const useDashboard = () => {
 		];
 
 		return allColumns;
-	}, [
-		contactHasName,
-		computeName,
-		handleCellHover,
-		stateBadgeColorMap,
-		canadianProvinceNames,
-		canadianProvinceAbbreviations,
-	]);
+	}, [contactHasName, computeName, handleCellHover]);
 
 	return {
 		form,
@@ -651,5 +589,10 @@ export const useDashboard = () => {
 		hasSearched,
 		handleResetSearch,
 		hoveredText,
+		handleSubmitSearch,
+		isMobile,
+		hasProblematicBrowser,
+		isSignedIn,
+		openSignIn,
 	};
 };
