+++ src/components/atoms/_svg/SearchIcon.tsx
@@ -1,0 +1,40 @@
+import { CommonIconProps, defaultPathClassName } from '@/utils';
+import { FC } from 'react';
+
+const SearchIcon: FC<CommonIconProps> = (props) => {
+	const {
+		width = '100%',
+		height = '100%',
+		className,
+		pathClassName = defaultPathClassName,
+	} = props;
+	return (
+		<svg
+			width={width}
+			height={height}
+			viewBox="0 0 20 21"
+			fill="none"
+			xmlns="http://www.w3.org/2000/svg"
+			aria-hidden="true"
+			className={className}
+		>
+			<path
+				d="M12 1C15.9278 1 19 3.96996 19 7.5C19 11.03 15.9278 14 12 14C8.07223 14 5 11.03 5 7.5C5 3.96996 8.07223 1 12 1Z"
+				stroke="#A0A0A0"
+				className={pathClassName}
+				strokeWidth="2"
+			/>
+			<line
+				x1="7.75258"
+				y1="11.6585"
+				x2="0.752577"
+				y2="19.6585"
+				stroke="#A0A0A0"
+				className={pathClassName}
+				strokeWidth="2"
+			/>
+		</svg>
+	);
+};
+
+export default SearchIcon;
