+++ src/components/organisms/_dialogs/IdentityDialog/IdentityDialog.tsx
@@ -14,6 +14,8 @@ import { Button } from '@/components/ui/button';
 
 import { urls } from '@/constants/urls';
 import { Typography } from '@/components/ui/typography';
+import { cn } from '@/utils';
+import SquareXIcon from '@/components/atoms/_svg/SquareX';
 
 export const IdentityDialog: FC<IdentityDialogProps> = (props) => {
 	const router = useRouter();
@@ -136,11 +138,9 @@ export const IdentityDialog: FC<IdentityDialogProps> = (props) => {
 					{/* Main content area - scrollable */}
 					<div className="flex-1 min-h-0 overflow-y-auto overflow-x-hidden bg-gray-50/30">
 						<div className="flex justify-center py-6">
-							<div className="w-full max-w-[1443.9px] px-0 mx-auto">
+							<div className="w-full max-w-[1444px] px-0 mx-auto">
 								{isPendingIdentities ? (
-									<div className="h-full flex items-center justify-center">
-										{/* Empty space during load - fade transition handles the visual feedback */}
-									</div>
+									<div className="h-full flex items-center justify-center"></div>
 								) : (
 									<div className="flex flex-col items-stretch justify-start">
 										{/* Show create form centered when no profiles exist */}
@@ -163,12 +163,14 @@ export const IdentityDialog: FC<IdentityDialogProps> = (props) => {
 											</div>
 										) : (
 											/* Show grid layout when profiles exist */
-											<div className="grid grid-cols-1 w-full lg:grid-cols-[652.4px_650.5px] lg:gap-[141px] lg:w-[1443.9px] lg:mx-auto">
+											<div className="grid grid-cols-1 w-full lg:grid-cols-[652px_651px] lg:gap-[141px] lg:w-[1444px] lg:mx-auto">
 												{/* Existing Profiles Section */}
-												<div style={{ opacity: showCreatePanel ? 0.26 : 1 }}>
+												<div
+													className={cn(showCreatePanel ? 'opacity-25' : 'opacity-100')}
+												>
 													<Typography
 														variant="h3"
-														className="!text-[18.77px] !leading-[22.1px] font-medium text-[#000000] mb-4 font-secondary"
+														className="!text-[19px] !leading-[22px] font-medium text-black mb-4 font-secondary"
 													>
 														Select User Profile
 													</Typography>
@@ -178,7 +180,7 @@ export const IdentityDialog: FC<IdentityDialogProps> = (props) => {
 															name="identityId"
 															render={({ field }) => (
 																<FormItem>
-																	<div className="box-border shrink-0 w-[652.4px] h-[326.75px] rounded-[8.83px] border-[2.21px] border-[#000000] overflow-hidden">
+																	<div className="box-border shrink-0 w-[653px] h-[327px] rounded-[9px] border-[2px] border-black overflow-hidden">
 																		<div className="w-full h-full overflow-y-auto overflow-x-hidden">
 																			<Table className="w-full !rounded-none">
 																				<TableBody>
@@ -194,19 +196,19 @@ export const IdentityDialog: FC<IdentityDialogProps> = (props) => {
 																								data-state={
 																									isSelected ? 'selected' : undefined
 																								}
-																								className="border-0 border-b border-[#000000] last:border-b-0 hover:!bg-transparent"
+																								className="border-0 border-b border-black last:border-b-0 hover:!bg-transparent"
 																							>
 																								<TableCell className="p-0">
-																									<div className="w-full h-[117.01px] flex flex-col justify-center gap-0 pl-4">
-																										<div className="font-primary font-normal text-[21.5px] text-black pl-1 mb-1">
+																									<div className="w-full h-[117px] flex flex-col justify-center gap-0 pl-4">
+																										<div className="font-primary font-normal text-[22px] text-black pl-1 mb-1">
 																											{identity.name}
 																										</div>
-																										<div className="w-[267.13px] h-[22.79px] bg-[#E8EFFF] border-[0.91px] border-[#000000] rounded-[7.29px] flex items-center px-2 overflow-hidden">
+																										<div className="w-[267px] h-[23px] bg-[#E8EFFF] border-[0.91px] border-[#000000] rounded-[7.29px] flex items-center px-2 overflow-hidden">
 																											<span className="font-secondary font-light text-[15.5px] text-[#000000] truncate">
 																												{identity.email}
 																											</span>
 																										</div>
-																										<div className="w-[267.13px] h-[22.79px] bg-[#E8EFFF] border-[0.91px] border-[#000000] rounded-[7.29px] flex items-center px-2 overflow-hidden mt-1">
+																										<div className="w-[267px] h-[23px] bg-[#E8EFFF] border-[0.91px] border-[#000000] rounded-[7.29px] flex items-center px-2 overflow-hidden mt-1">
 																											<span className="font-secondary font-light text-[15.5px] text-[#000000] truncate">
 																												{identity.website || 'No website'}
 																											</span>
@@ -224,17 +226,12 @@ export const IdentityDialog: FC<IdentityDialogProps> = (props) => {
 															)}
 														/>
 													</Form>
-													{/* Continue button 17px below left box */}
 													{!showCreatePanel && (
 														<div className="mt-[17px]">
 															<Button
 																onClick={handleAssignIdentity}
 																isLoading={isPendingAssignIdentity}
-																className="w-[652.4px] h-[43.05px] rounded-[8.83px] border-[1.1px] text-black"
-																style={{
-																	backgroundColor: 'rgba(93,171,104,0.49)',
-																	borderColor: '#5DAB68',
-																}}
+																className="w-[653px] h-[43px] rounded-[9px] border-1 text-black bg-[rgba(93,171,104,0.49)] border-[#5DAB68]"
 																disabled={!selectedIdentity}
 															>
 																Continue
@@ -253,7 +250,7 @@ export const IdentityDialog: FC<IdentityDialogProps> = (props) => {
 															<div className="flex items-center gap-2">
 																<Typography
 																	variant="h3"
-																	className="!text-[18.77px] !leading-[22.1px] font-medium text-[#000000] font-secondary"
+																	className="!text-[18px] !leading-[22px] font-medium text-black font-secondary"
 																>
 																	Create New Profile
 																</Typography>
@@ -292,27 +289,17 @@ export const IdentityDialog: FC<IdentityDialogProps> = (props) => {
 																onClick={(e) => e.stopPropagation()}
 																className="relative w-[651px]"
 															>
-																{/* Close (X) button positioned just above the top-right of the box */}
 																<button
 																	type="button"
 																	onClick={() => setShowCreatePanel(false)}
 																	aria-label="Close create profile"
 																	className="absolute -top-5 right-[8px] w-[13.05px] h-[13.05px] flex items-center justify-center rounded-none cursor-pointer bg-transparent"
 																>
-																	<svg
-																		width="13.05"
-																		height="13.05"
-																		viewBox="0 0 12 12"
-																		fill="none"
-																		xmlns="http://www.w3.org/2000/svg"
-																	>
-																		<path
-																			d="M1 1L11 11M11 1L1 11"
-																			stroke="#000000"
-																			strokeWidth="2"
-																			strokeLinecap="square"
-																		/>
-																	</svg>
+																	<SquareXIcon
+																		width="13px"
+																		height="13px"
+																		pathClassName="stroke-black fill-black"
+																	/>
 																</button>
 																<div className="p-0">
 																	<CreateIdentityPanel
